Refactor code structure to split up files
