+++ wsl-ssh-pageant.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27428.2015
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "wsl-ssh-pageant", "wsl-ssh-pageant.csproj", "{CC34838B-14B9-48D7-9B75-AEE9946D8EB9}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CC34838B-14B9-48D7-9B75-AEE9946D8EB9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CC34838B-14B9-48D7-9B75-AEE9946D8EB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CC34838B-14B9-48D7-9B75-AEE9946D8EB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CC34838B-14B9-48D7-9B75-AEE9946D8EB9}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {EB784A3D-6F60-4819-92EA-5887C29BDCB5}
+	EndGlobalSection
+EndGlobal
